Add files via upload
The file bengalurulocation contains some places of Bengaluru and the other file bustracking contains the original source code which u have to run. The input only takes places mentioned in the bengalurulocation file and nothing else since this is especially designed for the city of Bangalore. This is purely developed using java swing and awt :)